Move code from the JVM to common (naively, will not compile)
